+++ CHANGELOG.md
@@ -44,7 +44,7 @@
 - Album art: added new art type 'File (by TF)' which allows to assign art to specific nodes by TF, independently of foobar2000 art preferences. It may be used to load art based on folders or node names, instead of being track-based. e.g. For a genre/style tree, to show custom art for every genre (instead of art from the first track for such node); or show a custom image for every folder while using folder view. See [here](https://hydrogenaudio.org/index.php/topic,129076.msg1080070.html#msg1080070).
 - Album art: added menu entry to open the art cache folder (if used).
 - Album art: added new setting to show/hide the image gray frame border, which was forced on original script, at 'Album Art' tab (HTML options panel) or 'Image Frame border' (properties panel).
-- Album art: added new art styles: 'Star-Fill' (3), 'Star-Effect' (4), 'Star-Outline' (5).
+- Album art: added new art styles: 'Star-Fill' (3), 'Star-Effect' (4), 'Star-Outline' (5) and 'Heart' (6).
 - Album art: added new art flare hover effect. Setting can be found at 'Display' tab (HTML options panel) or 'Highlight Flare Image' (properties panel).
 - Statistics: added 3 new customizable statistics slots based on [X-power mean](https://en.wikipedia.org/wiki/Generalized_mean). Slots and labels can be modified easily at the HTML options panel or at 'Statistics Titleformat Custom-X (p-mean5)' and 'Statistics Titleformat Custom-X (p-mean5)' and 'Statistics Titleformat Custom labels' (properties panel). By default the panel includes custom expressions as example and sets the exponent to 3.
 - Readme: added help submenu, at settings menu, showing all available docs. See [here](https://hydrogenaudio.org/index.php/topic,129076.msg1078914.html#msg1078914).